requirements.txt
pyspark
requests
python-dateutil
google-cloud-bigquery
pandas
pytest
pytest-mock